tomato: 0.8.1
> Come on, can't you show a little mercy and compassion?

  * Fix pikaur build due to a base-devel missing;
  * Fix repose unable to import some packages.

+++ PKGBUILD
@@ -1,7 +1,7 @@
 # Maintainer : aji <43468383+aji-prod@users.noreply.github.com>
 pkgname=tomato
-pkgver=0.8.0
-pkgrel=2
+pkgver=0.8.1
+pkgrel=1
 pkgdesc="AUR local repository manager"
 arch=('any')
 url="https://github.com/aji-prod/tomato"
@@ -10,7 +10,7 @@ source=(
 	"${pkgname}-${pkgver}.tar.gz"::https://github.com/aji-prod/tomato/archive/v"${pkgver}".tar.gz
 )
 
-sha256sums=('525be6dd16ac5c332140a930f8752e64c987d8fb3a1bed60bce2c39b5bc5d8bf')
+sha256sums=('22f5b20c92920bb23b8accf702059dbb14fde6adc06cc92f296f5e4a878830a2')
 
 depends=(
 	'docker'
@@ -21,7 +21,16 @@ package() {
 	install -Dm755 tomato "$pkgdir/usr/bin/tomato"
 	install -Dm644 tomato.conf "$pkgdir/etc/tomato.conf"
 	for dockfile in $(cd ./docker && ls -A ./*); do
-		install -Dm644 docker/$dockfile "$pkgdir/usr/share/$pkgname/$dockfile"
+		if test -f docker/$dokerfile;
+		then
+			install -Dm644 docker/$dockfile "$pkgdir/usr/share/$pkgname/$dockfile"
+		fi
+	done
+	for dockfile in $(cd ./docker && ls -A ./patch/*); do
+		if test -f docker/patch/$dokerfile;
+		then
+			install -Dm644 docker/patch/$dockfile "$pkgdir/usr/share/$pkgname/patch/$dockfile"
+		fi
 	done
 	for hookfile in $(cd ./pacman && ls ./*); do
 		install -Dm644 pacman/$hookfile "$pkgdir/usr/share/libalpm/hooks/$hookfile"
